+++ app/Resources/views/AddObservation/addObservation.html.twig
@@ -2,7 +2,7 @@
 {% block content %}
     <div class="row">
         <h1>Faire une observation</h1>
-        {% include 'default/form.html.twig' %}
+        {% include 'AddObservation/form.html.twig' %}
     </div>
 
 {% endblock %}
