+++ src/AppBundle/Services/Notification/Notification.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace AppBundle\Services;
+namespace AppBundle\Services\Notification;
 
 
 use Doctrine\ORM\EntityManagerInterface;
